feat: change idx from uint64_t in int64_t

+++ pocket-daos/src/dao-user.cpp
@@ -46,12 +46,12 @@ user::opt dao_user::get()
         for(auto&& row : **opt_rs)
         {
             user ret {
-                    .id = static_cast<uint64_t>(row["id"].to_integer()),
+                    .id = row["id"].to_integer(),
                     .name = row["name"].to_text(),
                     .email = row["email"].to_text(),
                     .passwd = row["passwd"].to_text(),
                     .status = user::stat(row["passwd"].to_integer()),
-                    .timestamp_last_update = static_cast<uint64_t>(row["timestamp_last_update"].to_integer())
+                    .timestamp_last_update = row["timestamp_last_update"].to_integer()
             };
             return ret;
         }
@@ -66,12 +66,12 @@ user::opt dao_user::login(const string& email, const string& passwd)
         for(auto&& row : **opt_rs)
         {
             user ret {
-                    .id = static_cast<uint64_t>(row["id"].to_integer()),
+                    .id = row["id"].to_integer(),
                     .name = row["name"].to_text(),
                     .email = row["email"].to_text(),
                     .passwd = row["passwd"].to_text(),
                     .status = user::stat(row["passwd"].to_integer()),
-                    .timestamp_last_update = static_cast<uint64_t>(row["timestamp_last_update"].to_integer())
+                    .timestamp_last_update = row["timestamp_last_update"].to_integer()
             };
             return ret;
         }
